Added BindToChannel features
Added the ability to bind the bot to a channel.  When this is
implemented, the bot will only listen to, and respond in, the channel
specified in the config.properties file.

Cleaned up bot responses for when only 1 card match is found.  Bot will
no longer provide a list of matched cards when only 1 card is found.